requirements.txt
discord.py==2.4.0
asyncpg==0.29.0
python-dotenv==1.0.1
uvloop==0.20.0; sys_platform != 'win32'
